requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-watson==1.6.3
mysqlclient==2.2.5
psycopg2==2.9.10
python-dotenv==1.0.1
redis==5.2.0
sqlparse==0.5.1
tzdata==2024.2
